Move `RequestUtils` trait and `redirect()` fn to `util::request_helpers` module
