+++ src/controllers.rs
@@ -18,46 +18,10 @@ mod prelude {
     pub use http::{header, request::Parts, StatusCode};
 
     pub use crate::app::AppState;
-    use crate::controllers::util::RequestPartsExt;
     pub use crate::middleware::app::RequestApp;
     pub use crate::tasks::spawn_blocking;
     pub use crate::util::errors::{AppResult, BoxedAppError};
-    pub use crate::util::BytesRequest;
-    use indexmap::IndexMap;
-
-    pub fn redirect(url: String) -> Response {
-        (StatusCode::FOUND, [(header::LOCATION, url)]).into_response()
-    }
-
-    pub trait RequestUtils {
-        fn query(&self) -> IndexMap<String, String>;
-        fn wants_json(&self) -> bool;
-        fn query_with_params(&self, params: IndexMap<String, String>) -> String;
-    }
-
-    impl<T: RequestPartsExt> RequestUtils for T {
-        fn query(&self) -> IndexMap<String, String> {
-            url::form_urlencoded::parse(self.uri().query().unwrap_or("").as_bytes())
-                .into_owned()
-                .collect()
-        }
-
-        fn wants_json(&self) -> bool {
-            self.headers()
-                .get_all(header::ACCEPT)
-                .iter()
-                .any(|val| val.to_str().unwrap_or_default().contains("json"))
-        }
-
-        fn query_with_params(&self, new_params: IndexMap<String, String>) -> String {
-            let mut params = self.query();
-            params.extend(new_params);
-            let query_string = url::form_urlencoded::Serializer::new(String::new())
-                .extend_pairs(params)
-                .finish();
-            format!("?{query_string}")
-        }
-    }
+    pub use crate::util::{redirect, BytesRequest, RequestUtils};
 }
 
 pub mod helpers;
