+++ src/util/request_helpers.rs
@@ -1,5 +1,8 @@
+use crate::controllers::util::RequestPartsExt;
+use axum::response::{IntoResponse, Response};
 use http::header::AsHeaderName;
-use http::HeaderMap;
+use http::{header, HeaderMap, StatusCode};
+use indexmap::IndexMap;
 
 pub trait HeaderMapExt {
     /// Returns the value of the request header, or an empty slice if it is not
@@ -16,5 +19,39 @@ impl HeaderMapExt for HeaderMap {
         self.get(key)
             .map(|value| value.to_str().unwrap_or_default())
             .unwrap_or_default()
+    }
+}
+
+pub fn redirect(url: String) -> Response {
+    (StatusCode::FOUND, [(header::LOCATION, url)]).into_response()
+}
+
+pub trait RequestUtils {
+    fn query(&self) -> IndexMap<String, String>;
+    fn wants_json(&self) -> bool;
+    fn query_with_params(&self, params: IndexMap<String, String>) -> String;
+}
+
+impl<T: RequestPartsExt> RequestUtils for T {
+    fn query(&self) -> IndexMap<String, String> {
+        url::form_urlencoded::parse(self.uri().query().unwrap_or("").as_bytes())
+            .into_owned()
+            .collect()
+    }
+
+    fn wants_json(&self) -> bool {
+        self.headers()
+            .get_all(header::ACCEPT)
+            .iter()
+            .any(|val| val.to_str().unwrap_or_default().contains("json"))
+    }
+
+    fn query_with_params(&self, new_params: IndexMap<String, String>) -> String {
+        let mut params = self.query();
+        params.extend(new_params);
+        let query_string = url::form_urlencoded::Serializer::new(String::new())
+            .extend_pairs(params)
+            .finish();
+        format!("?{query_string}")
     }
 }
